requirements.txt
asgiref==3.5.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
cssselect==1.1.0
cssutils==2.4.0
dj-database-url==0.5.0
Django==4.0.4
django-ckeditor==6.3.2
django-cloudinary-storage==0.3.0
django-heroku==0.0.0
django-js-asset==2.0.0
gunicorn==20.1.0
idna==3.3
lxml==4.8.0
Pillow==9.1.0
premailer==3.10.0
psycopg2-binary==2.9.3
python-decouple==3.6
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.0.0
yagmail==0.15.277
